Fix for bug 14 QB failed to load data.
There were numbers in the Num column that had non-numeric characters.

I will correct so that non-numeric characters will be ignored in future
versions. It is expected that these are integers only, so I am keeping
only the digits and removing all other characters.

If you want the hyphen or any other characters (like decimal points or
commas), then this data will have to be stored as text. LGL will allow
check numbers to be stored as strings if that is your preference. I can
restrict the acceptable characters (I doubt that anyone uses a grave in
a check number. A hyphen is a possibility, so I can allow hyphens).